Narrow workerId and datacenterId from Long to Int
Snowflake and Flake only use small worker/datacenter values — typically
≤ 5 bits for each in practice, and in theory capped at 31 bits to fit in
a positive Int. Long was wider than necessary.

- FlakeIdGenerator: `workerId: Long` → `workerId: Int`, `datacenterId: Long`
  → `datacenterId: Int`, and the matching public `maxWorkerId`/`maxDatacenterId`
  are now `Int`. Final bit-packing converts via `.toLong()`.
- FlakeIdGenerator: enforce `workerIdBits in 1..31` so `(1 shl workerIdBits) - 1`
  stays a positive Int. 31 bits allows up to ~2.1B workers, far beyond any
  realistic deployment.
- SnowflakeIdGenerator: mirror the parent's new types.
- Tests: local `maxWorkerId`/`maxDatacenterId` helpers become `Int`; bit
  extraction converts the Long id slice via `.toInt()` before masking.

Breaking change. Positional callers that passed `Long` literals or variables
must switch to `Int`. Still part of the 2.0.0 release.

+++ src/main/kotlin/io/github/dornol/idkit/flake/SnowflakeIdGenerator.kt
@@ -15,8 +15,8 @@ import java.time.Instant
  * @see <a href="https://github.com/twitter/snowflake/tree/snowflake-2010">Twitter Snowflake</a>
  */
 class SnowflakeIdGenerator(
-    workerId: Long,
-    datacenterId: Long,
+    workerId: Int,
+    datacenterId: Int,
     epochStart: Instant = Instant.EPOCH,
 ) : FlakeIdGenerator(
     timestampBits = 41,
